Update file siswa to daftar in pengumuman admin
Previously it's just one column for each one file
with preview and download button (e.g. File Prestasi
has preview and download, and so on for File Keahlian).

Now it's like page pendidikan in admin which is having
one button to show modal that has all data jurusan
related to pendidikan.